Refactor TableMetadataManager
