+++ core/src/main/java/com/scalar/db/storage/common/checker/OperationChecker.java
@@ -11,8 +11,8 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.Key;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
