+++ core/src/main/java/com/scalar/db/storage/cosmos/SelectStatementHandler.java
@@ -12,8 +12,8 @@ import com.scalar.db.api.Scan;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.IntStream;
