+++ core/src/main/java/com/scalar/db/storage/dynamo/BatchHandler.java
@@ -8,7 +8,7 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.exception.storage.NoMutationException;
 import com.scalar.db.exception.storage.RetriableExecutionException;
-import com.scalar.db.storage.common.TableMetadataManager;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
