+++ core/src/main/java/com/scalar/db/storage/dynamo/Dynamo.java
@@ -13,9 +13,9 @@ import com.scalar.db.api.Scan;
 import com.scalar.db.api.Scanner;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.common.checker.OperationChecker;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import java.net.URI;
 import java.util.List;
 import java.util.Map;
@@ -65,7 +65,9 @@ public class Dynamo implements DistributedStorage {
     namespace = Optional.empty();
     tableName = Optional.empty();
 
-    metadataManager = new TableMetadataManager(new DynamoAdmin(client, config), config);
+    metadataManager =
+        new TableMetadataManager(
+            new DynamoAdmin(client, config), config.getTableMetadataCacheExpirationTimeSecs());
     operationChecker = new OperationChecker(metadataManager);
 
     selectStatementHandler = new SelectStatementHandler(client, metadataManager);
