+++ core/src/main/java/com/scalar/db/storage/dynamo/SelectStatementHandler.java
@@ -12,10 +12,10 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.Key;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.dynamo.bytes.BytesUtils;
 import com.scalar.db.storage.dynamo.bytes.KeyBytesEncoder;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Collections;
