+++ core/src/main/java/com/scalar/db/transaction/jdbc/JdbcTransactionManager.java
@@ -7,7 +7,6 @@ import com.scalar.db.api.Isolation;
 import com.scalar.db.api.SerializableStrategy;
 import com.scalar.db.api.TransactionState;
 import com.scalar.db.exception.transaction.TransactionException;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.common.checker.OperationChecker;
 import com.scalar.db.storage.jdbc.JdbcConfig;
 import com.scalar.db.storage.jdbc.JdbcDatabaseAdmin;
@@ -15,6 +14,7 @@ import com.scalar.db.storage.jdbc.JdbcService;
 import com.scalar.db.storage.jdbc.JdbcUtils;
 import com.scalar.db.storage.jdbc.RdbEngine;
 import com.scalar.db.storage.jdbc.query.QueryBuilder;
+import com.scalar.db.util.TableMetadataManager;
 import java.sql.SQLException;
 import java.util.Optional;
 import java.util.UUID;
@@ -41,7 +41,9 @@ public class JdbcTransactionManager implements DistributedTransactionManager {
 
     tableMetadataDataSource = JdbcUtils.initDataSourceForTableMetadata(config);
     TableMetadataManager tableMetadataManager =
-        new TableMetadataManager(new JdbcDatabaseAdmin(tableMetadataDataSource, config), config);
+        new TableMetadataManager(
+            new JdbcDatabaseAdmin(tableMetadataDataSource, config),
+            config.getTableMetadataCacheExpirationTimeSecs());
 
     OperationChecker operationChecker = new OperationChecker(tableMetadataManager);
     QueryBuilder queryBuilder = new QueryBuilder(rdbEngine);
