+++ core/src/main/java/com/scalar/db/transaction/rpc/GrpcTwoPhaseCommitTransactionOnBidirectionalStream.java
@@ -33,10 +33,10 @@ import com.scalar.db.rpc.TwoPhaseCommitTransactionRequest.ValidateRequest;
 import com.scalar.db.rpc.TwoPhaseCommitTransactionResponse;
 import com.scalar.db.rpc.TwoPhaseCommitTransactionResponse.Error.ErrorCode;
 import com.scalar.db.rpc.TwoPhaseCommitTransactionResponse.GetResponse;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.rpc.GrpcConfig;
 import com.scalar.db.util.ProtoUtils;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import com.scalar.db.util.retry.ServiceTemporaryUnavailableException;
 import io.grpc.Status.Code;
 import io.grpc.StatusRuntimeException;
