+++ core/src/test/java/com/scalar/db/storage/common/checker/OperationCheckerTest.java
@@ -25,8 +25,8 @@ import com.scalar.db.io.IntValue;
 import com.scalar.db.io.Key;
 import com.scalar.db.io.TextValue;
 import com.scalar.db.io.Value;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.util.ScalarDbUtils;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
