+++ core/src/test/java/com/scalar/db/storage/cosmos/PutStatementHandlerTest.java
@@ -27,7 +27,7 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.NoMutationException;
 import com.scalar.db.exception.storage.RetriableExecutionException;
 import com.scalar.db.io.Key;
-import com.scalar.db.storage.common.TableMetadataManager;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.List;
