+++ core/src/test/java/com/scalar/db/storage/dynamo/DeleteStatementHandlerTest.java
@@ -16,7 +16,7 @@ import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.exception.storage.NoMutationException;
 import com.scalar.db.io.Key;
-import com.scalar.db.storage.common.TableMetadataManager;
+import com.scalar.db.util.TableMetadataManager;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import org.junit.Before;
