+++ core/src/test/java/com/scalar/db/storage/dynamo/SelectStatementHandlerTest.java
@@ -17,9 +17,9 @@ import com.scalar.db.api.Scan.Ordering.Order;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.Key;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.dynamo.bytes.BytesUtils;
 import com.scalar.db.storage.dynamo.bytes.KeyBytesEncoder;
+import com.scalar.db.util.TableMetadataManager;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.Collections;
