+++ core/src/test/java/com/scalar/db/storage/rpc/GrpcStorageTest.java
@@ -15,7 +15,7 @@ import com.scalar.db.exception.storage.NoMutationException;
 import com.scalar.db.io.Key;
 import com.scalar.db.rpc.DistributedStorageGrpc;
 import com.scalar.db.rpc.GetResponse;
-import com.scalar.db.storage.common.TableMetadataManager;
+import com.scalar.db.util.TableMetadataManager;
 import io.grpc.Status;
 import java.util.Arrays;
 import java.util.List;
