+++ core/src/test/java/com/scalar/db/transaction/rpc/GrpcTwoPhaseCommitTransactionManagerTest.java
@@ -12,8 +12,8 @@ import com.scalar.db.rpc.AbortResponse;
 import com.scalar.db.rpc.GetTransactionStateResponse;
 import com.scalar.db.rpc.TransactionState;
 import com.scalar.db.rpc.TwoPhaseCommitTransactionGrpc;
-import com.scalar.db.storage.common.TableMetadataManager;
 import com.scalar.db.storage.rpc.GrpcConfig;
+import com.scalar.db.util.TableMetadataManager;
 import io.grpc.Status;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.Before;
